scripts/requirements.txt
# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# YouTube
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0

# Google Gemini API (LLM quote extraction)
google-generativeai>=0.5.0

# Environment variable management
python-dotenv>=1.0.0

# Fuzzy string matching (film title and name matching)
thefuzz>=0.22.0
python-Levenshtein>=0.25.0

# Progress bars
tqdm>=4.66.0

# Rate limiting
ratelimit>=2.2.1

# Data validation
jsonschema>=4.20.0
